feat(packages/esbuild-config): build 설정을 위한 패키지 생성
