+++ packages/core/build.js
@@ -1,39 +1,4 @@
-import esbuild from "esbuild";
+import build from "@puzzlepop2/esbuild-config";
 import pkg from "./package.json" assert { type: "json" };
 
-const dev = process.argv.includes("--dev");
-const minify = !dev;
-
-const watch = process.argv.includes("--watch");
-
-const external = Object.keys({
-  ...pkg.dependencies,
-  ...pkg.peerDependencies,
-});
-
-const baseConfig = {
-  entryPoints: ["src/index.ts"],
-  bundle: true,
-  minify,
-  sourcemap: true,
-  outdir: "dist",
-  watch,
-  external,
-};
-
-Promise.all([
-  esbuild.build({
-    ...baseConfig,
-    format: "esm",
-  }),
-  esbuild.build({
-    ...baseConfig,
-    format: "cjs",
-    outExtension: {
-      ".js": ".cjs",
-    },
-  }),
-]).catch(() => {
-  console.error("Build failed");
-  process.exit(1);
-});
+build({ pkg });
