+++ packages/core/package.json
@@ -22,13 +22,13 @@
   },
   "devDependencies": {
     "@puzzlepop2/ts-config": "*",
+    "@puzzlepop2/esbuild-config": "*",
     "@babel/core": "^7.26.0",
     "@babel/preset-env": "^7.26.0",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.10.7",
     "@typescript-eslint/eslint-plugin": "^6.5.0",
     "@typescript-eslint/parser": "^6.5.0",
-    "esbuild": "0.16.17",
     "eslint": "^8.48.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.28.1",
