+++ packages/esbuild-config/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@puzzlepop2/esbuild-config",
+  "version": "0.0.0",
+  "private": true,
+  "main": "index.js",
+  "files": [
+    "index.js"
+  ],
+  "scripts": {
+    "clean": "rm -rf node_modules",
+    "build": ""
+  },
+  "dependencies": {
+    "esbuild": "0.16.17"
+  }
+}
